SONARKT-226 Re-enable S125 (#211)
The old approach to detecting commented-out code was trying to parse comments using the Kotlin compiler and if
successful raising an issue. This approach was quite slow and resulted in S125 being disabled silently for a
while. To re-enable this rule, this new implementation makes use of the sonar-analyzer-commons-recognizers
library as a pre-filter, which only does basic pattern matching but is significantly more performant. Only
if the prefiltering concludes that it could be code, is a comment parsed (avoiding more FPs).